The remains of the last great Fruit Empire
After the version control war GitHub arose victorious from the ashes of
the battlefield. However, the code lived on, in a fragmented state. The
code was able to survive the years of not being backed upped and
eventually came back in the form of a beautiful thing. Now it's almost
like a blank slate, something letting us all start out new.